requirements.txt
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
psutil==6.0.0
requests==2.32.3
click==8.1.7
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
pgvector==0.3.6
pytest==8.3.3
pytest-cov==5.0.0
ruff==0.12.3